% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef_techniques}
\alias{coef_techniques}
\title{coef_techniques}
\usage{
coef_techniques(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\value{
a matrix
}
\description{
coef_techniques
}
\examples{

}
