% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{identite}
\alias{identite}
\title{Identite}
\usage{
identite(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
a matrix
}
\description{
Identite
}
\examples{
A <- matrix(data = 1:16, nrow = 4)
identite(A)

}
